requirements.txt
osonbot
google-genai